pyproject.toml
[project]
name = "bisect-bot"
version = "1.0.0"
description = "A GitHub App that performs git bisection when triggered by issue comments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # GitHub integration
    "PyGithub>=2.1.1",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.7",
    # HTTP client
    "httpx>=0.26.0",
    # Docker SDK
    "docker>=7.0.0",
    # Docker runner for spawning isolated containers
    "dockerrun @ git+https://github.com/bobrenjc93/dockerrun.git",
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    # Environment and configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

